test: add count broken links function test
